Add files via upload
# Smart-Water-Management-System
Designed by our team , this Water management system is a solution which is low cost and easy to maintain. 
It has been designed keeping minimalism in perspective and affordability at priority so it can be made accessible to mass. 
Considering the design of this Project, one can reconfigure it according to his/her needs and implement it for themselves. 
I/Our team promotes open environment where everyone can design, customized in accordance of their needs. 
Feel free to use this idea as your practical project.